CNN_PyTorch/requirements.txt
torch
torchvision
Pillow
tqdm
tk
